feat: create basic collection view
