+++ AppCore/Sources/Screens/CreateRecipe/Models/StepTwoState.swift
@@ -36,18 +36,3 @@ struct StepTwoState: Equatable {
         areIngredientsValid = ingredients.allSatisfy(\.isValid)
     }
 }
-
-struct NewIngredient: Equatable {
-    let id: String
-    var name: String
-    var amount: Int?
-    var unit: IngredientUnit
-
-    var isValid: Bool {
-        !name.isEmpty
-    }
-
-    static func makeNewIngredient() -> NewIngredient {
-        .init(id: UUID().uuidString, name: "", amount: nil, unit: .gram)
-    }
-}
